pyproject.toml
[project]
name = "mstar-atr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "hydra-core>=1.3.0",
    "hydra-zen>=0.16.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "omegaconf>=2.3.0",
    "pillow>=10.0.0",
    "scikit-learn>=1.3.0",
    "seaborn>=0.12.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
]
